### 2025-10-03 #9134 @editorialbot
[PRE REVIEW]: ect: A Python Package for the Euler Characteristic Transform
**Submitting author:** @yemeen (<a href="http://orcid.org/0009-0009-6667-0991">Yemeen Ayub</a>)
**Repository:** https://github.com/MunchLab/ect
**Branch with paper.md** (empty if default branch): JOSS-Paper
**Version:** 1.2.0
