### 2025-11-24 #9134 @prashjha
Hi @edeno, @LuisScoccola, @robbievanleeuwen, @mtsch, @rrrlw, @shaelebrown, I am looking for two reviewers for the above JOSS submission, and this may be in your domain. We also encourage authors who recently published for JOSS to contribute to the review process.

Would you be interested in reviewing this? You can find pdf at this link [pdf link](https://github.com/openjournals/joss-papers/blob/joss.09134/joss.09134/10.21105.joss.09134.pdf), and the corresponding software is at this [link](https://github.com/MunchLab/ect).

This [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
