### 2025-12-10 #9134 @yemeen
@raphaelreinauer says about 2-4 weeks https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#time-expectations
