### 2025-12-26 MunchLab/ect#71 @corybrunson
acknowledge all authors' software contributions
Acknowledgments to the fourth author (Chitwood) are easily found by searching the repo or documentation but were not clear from the readme. While i've checked this item—contribution and authorship—as completed (it won't affect my review), i suggest for transparency that all authors be mentioned there.

_Part of [this _JOSS_ review](https://github.com/openjournals/joss-reviews/issues/9691)._
