### 2025-12-26 MunchLab/ect#72 @corybrunson
identify specific other ECT software
The statement of need in the manuscript alludes to other Python implementations of the Euler characteristic transform without citing them. Per the "state of the field" criterion, could the manuscript describe the limitations of at least the most popular of these, or cite a previous software review if one exists? (I see this as important both to justify the improvement claim and to inform new users of their options.)

_Part of [this _JOSS_ review](https://github.com/openjournals/joss-reviews/issues/9691)._
