### 2025-12-29 MunchLab/ect#73 @corybrunson
reproduce manuscript example & debug Tutorial 3.2
I'm working through the (clear and helpful) documentation. Though i installed `ect` from PyPI using pip3, i encountered some bugs in [3.2. Tutorial](https://munchlab.github.io/ect/notebooks/Tutorial-ExactECT.html), possibly due to breaking changes in more recent versions of the package:

* The line `from ect import ECT, EmbeddedComplex, create_example_graph` fails with `ImportError: cannot import name 'create_example_graph' from 'ect' (/Library/Frameworks/Python.framework/Versions/3.14/lib/python3.14/site-packages/ect/__init__.py)`.
* The `add_node` lines are missing brackets, e.g. `G.add_node('A', 0,0)` needs to be `G.add_node('A', [0,0])`.
* Later commands using e.g. `plot_angle_circle` and `rescale_to_unit_disk` fail as these names/attributes are not defined.

I apologize if these should be straightforward to resolve; Python is not my first language.

So that readers can reproduce the example in the manuscript, it would be helpful to link to this tutorial, or else include minimal source code to generate the figures.

_Part of [this _JOSS_ review](https://github.com/openjournals/joss-reviews/issues/9691)._
