### 2025-12-29 MunchLab/ect#74 @corybrunson
match citation usage & syntax
The citations in the manuscript section "The Euler Characteristic Transform" are formatted parenthetically but treated as in-text; either the usage or the syntax should be fixed. For example, `(See e.g. [@Hatcher])` should instead be `(See e.g. @Hatcher)` or perhaps just `[@Hatcher]`. (Here i'm referring to [this section](https://pandoc.org/MANUAL.html#extension-citations) of the Pandoc manual.)

_Part of [this _JOSS_ review](https://github.com/openjournals/joss-reviews/issues/9691)._
