### 2025-12-29 MunchLab/ect#75 @corybrunson
statement of need in package documentation
_JOSS_ expects the package documentation itself to include a statement of need. Based on previous conversations, this need not be a full section as in the manuscript but could be a sentence or two in [Getting Started](https://munchlab.github.io/ect/). Maybe it's already there in another section; if so, please point me to it and accept my apology!

_Part of [this _JOSS_ review](https://github.com/openjournals/joss-reviews/issues/9691)._
