### 2025-12-30 MunchLab/ect#80 @raphaelreinauer
Confirm Numba performance claims with benchmarks
Related to: https://github.com/MunchLab/ect/issues/72#issuecomment-3696712933

Per the JOSS review checklist:
> If there are any performance claims of the software, have they been confirmed?

The paper has performance claims (e.g., “significant speedups over naive Python implementations”), which are specifically attributed to Numba JIT compilation, but the paper and the package don’t currently contain benchmarks comparing the JIT implementation against a non-JIT baseline. Since the repo already contains benchmarks, it would be helpful to add a small benchmark that includes at least one baseline comparison (e.g., Numba enabled vs disabled) to support the speedup claim.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/9691).
