### 2025-12-30 MunchLab/ect#81 @raphaelreinauer
[Non-blocking] Minor code quality issues: unused imports and type annotations
Per the JOSS review checklist:
> **Functionality:** Have the functional claims of the software been confirmed?

All functionality works correctly (101 tests pass), but static analysis show some issue:

1. Unused imports:
```
> ruff check src/ect/

src/ect/embed_complex.py:12-14: F401 - unused imports: point_in_polygon, validate_face_embedding, validate_edge_embedding
src/ect/sect.py:5: F401 - unused import: Union
```

2. Type annotation issues:
```
> mypy src/ect/ --ignore-missing-imports
```
- 8 implicit Optional violations (PEP 484) in validation/rules.py and validation/base.py
- Type mismatches in directions.py (lines 117, 126, 131, 147, 157, 164)

These issues are non-blocking but would be great to address.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/9691).
