### 2025-12-30 MunchLab/ect#82 @raphaelreinauer
Clarify theta argument validation error message
In `src/ect/ect.py:84-87`:
```
        if theta is not None and graph_dim != 2:
            raise ValueError(
                "Theta must be provided for 2D graphs. "
                "Use 'directions' or 'num_dirs' to specify directions."
            )
```

The condition rejects theta for non-2D graphs, but the message says theta "must be provided for 2D graphs", which is the opposite.

Suggested fix:
```
raise ValueError(
    "theta is only supported for 2D graphs. "
    "Use 'directions' or 'num_dirs' for higher dimensions."
)
```

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/9691).
