### 2025-12-30 MunchLab/ect#83 @raphaelreinauer
Clarify target audience in Statement of Need
Per the JOSS review checklist:
> **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other
work?

The current "Statement of Need" states the problem but the target audience is only implicite in "modern research datasets".

Consider explicitly adding the target audience.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/9691).
