### 2025-12-30 MunchLab/ect#84 @raphaelreinauer
Minor typos and grammar fixes in paper
Per the JOSS review checklist:
> **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

There are some minor grammar issues/typos in the paper:

Summary section:
- "scientific package" → "scientific packages"

Statement of Need section:
- "addressed" → "addresses"
- "complimented" → "complemented"
- "Tranforms" → "Transforms"

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/9691).
