### 2025-12-30 MunchLab/ect#85 @raphaelreinauer
Add citation for ECT completeness claim
Per the JOSS review checklist:
> Is the list of references complete, and is everything cited appropriately that should be cited?

The Summary states the ECT "provid[es] a complete topological summary" without a direct citation.
This refers to the injectivity result. Consider adding a citation here, e.g., Turner2014 or Munch2025.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/9691).
