### 2025-12-30 #9691 @raphaelreinauer
@prashjha I've completed my first pass and opened seven minor issues on the repo.

@yemeen Thanks for the submission! Let me know if you have any questions about the issues. Looking forward to your updates.
