### 2025-12-30 MunchLab/ect#86 @raphaelreinauer
Add citations for SECT and DECT
Per the JOSS review checklist:
> Is the list of references complete, and is everything cited appropriately that should be cited?

The Statement of Need mentions "the ECT, SECT, and the DECT" but neither SECT nor DECT are cited in the paper.

Please add relevant citations.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/9691).
