### 2025-12-30 MunchLab/ect#87 @raphaelreinauer
Consider adding notebook execution to CI
Related to: https://github.com/MunchLab/ect/issues/73

Per the JOSS review checklist:
>*Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

Consider adding notebook execution to automated tests to detect issues, e.g., Tutorial-ExactECT.ipynb currently fails to execute due to API changes.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/9691).
