### 2026-01-21 d085956
remove heuristic
