### 2026-01-22 e39c843
Yet more documentation cleanup
