+++ src/ect/embed_complex.py
@@ -22,7 +22,7 @@ TRANSFORM_TYPES = ["pca"]
 
 class EmbeddedComplex(nx.Graph):
     """
-    A unified class to represent an embedded cell complex with cells of arbitrary dimension.
+    A unified class to represent an embedded cell complex with cells of arbitrary dimension. 
 
     This combines the functionality of EmbeddedGraph and EmbeddedCW, supporting:
     - 0-cells (vertices) with embedded coordinates
@@ -53,6 +53,17 @@ class EmbeddedComplex(nx.Graph):
     """
 
     def __init__(self, validate_embedding=False, embedding_tol=1e-10):
+        """
+        Initialize an EmbeddedComplex instance.
+
+        Args:
+            validate_embedding (bool, optional): If True, automatically validate embedding properties when adding cells. Defaults to False.
+            embedding_tol (float, optional): Tolerance for geometric validation. Defaults to 1e-10.
+
+        Notes:
+            - The complex supports arbitrary-dimensional cells (vertices, edges, faces, etc.).
+            - Embedding validation can be enabled or disabled at initialization or later.
+        """
         super().__init__()
         self._node_list = []
         self._node_to_index = {}
@@ -74,36 +85,74 @@ class EmbeddedComplex(nx.Graph):
 
     @property
     def coord_matrix(self):
-        """Return the N x D coordinate matrix"""
+        """
+        Get the coordinate matrix of the embedded complex.
+
+        Returns:
+            np.ndarray: A matrix of shape :math:`(N, D)` where :math:`N` is the number of nodes (vertices)
+            and :math:`D` is the dimension of the embedding. Each row corresponds to the coordinates
+            of a node in the order given by `node_list`.
+
+        Notes:
+            - If no nodes have been added, returns an empty array of shape :math:`(0, 0)`.
+            - The coordinate matrix is updated automatically as nodes are added or modified.
+            - The order of rows matches the order of node identifiers in `node_list`.
+        """
         if self._coord_matrix is None:
             return np.empty((0, 0))
         return self._coord_matrix
 
     @property
     def dim(self):
-        """Return the dimension of the embedded coordinates"""
+        """
+        Get the dimension of the embedded coordinates.
+
+        Returns:
+            int: The number of dimensions (D) for the coordinates of each node. Returns 0 if no nodes exist.
+        """
         if self._coord_matrix is None:
             return 0
         return self._coord_matrix.shape[1]
 
     @property
     def node_list(self):
-        """Return ordered list of node names."""
+        """
+        Get the ordered list of node names in the complex.
+
+        Returns:
+            list: List of node identifiers in the order corresponding to the rows of the coordinate matrix.
+        """
         return self._node_list
 
     @property
     def node_to_index(self):
-        """Return a dictionary mapping node ids to their index in the coord_matrix"""
+        """
+        Get a mapping from node identifiers to their row index in the coordinate matrix.
+
+        Returns:
+            dict: Dictionary mapping node ids to their index in the coordinate matrix.
+        """
         return self._node_to_index
 
     @property
     def position_dict(self):
-        """Return a dictionary mapping node ids to their coordinates"""
+        """
+        Get a dictionary mapping node ids to their coordinates.
+
+        Returns:
+            dict: Dictionary where keys are node ids and values are coordinate arrays for each node.
+        """
         return {node: self._coord_matrix[i] for i, node in enumerate(self._node_list)}
 
     @property
     def edge_indices(self):
-        """Return edges as array of index pairs"""
+        """
+        Get the edges of the complex as an array of index pairs.
+
+        Returns:
+            np.ndarray: Array of shape (E, 2) where each row contains the indices of the two nodes forming an edge.
[…]
