### 2026-01-22 8e61acd
Documentation expansion and/or cleanup
