+++ src/ect/embed_complex.py
@@ -1,5 +1,6 @@
 from collections import defaultdict
 from typing import Dict, List, Tuple, Optional, Union
+import functools
 
 import networkx as nx
 import numpy as np
@@ -22,9 +23,10 @@ TRANSFORM_TYPES = ["pca"]
 
 class EmbeddedComplex(nx.Graph):
     """
-    A unified class to represent an embedded cell complex with cells of arbitrary dimension. 
+    A unified class to represent an embedded cell complex with cells of arbitrary dimension.
 
-    This combines the functionality of EmbeddedGraph and EmbeddedCW, supporting:
+    This combines the functionality of :class:`EmbeddedGraph` and :class:`EmbeddedCW`, supporting:
+
     - 0-cells (vertices) with embedded coordinates
     - 1-cells (edges)
     - k-cells for k >= 2 (faces, volumes, etc.)
@@ -34,22 +36,6 @@ class EmbeddedComplex(nx.Graph):
             when adding cells. Default: False
         embedding_tol (float): Tolerance for geometric validation. Default: 1e-10
 
-    Attributes:
-        coord_matrix : np.ndarray
-            A matrix of embedded coordinates for each vertex
-        node_list : list
-            A list of node names
-        node_to_index : dict
-            A dictionary mapping node ids to their index in the coord_matrix
-        dim : int
-            The dimension of the embedded coordinates
-        cells : dict
-            Dictionary mapping dimension k to list of k-cells, where each k-cell
-            is represented as a tuple of vertex indices
-        validate_embedding : bool
-            Whether to automatically validate embedding properties
-        embedding_tol : float
-            Tolerance for geometric validation
     """
 
     def __init__(self, validate_embedding=False, embedding_tol=1e-10):
@@ -91,12 +77,12 @@ class EmbeddedComplex(nx.Graph):
         Returns:
             np.ndarray: A matrix of shape :math:`(N, D)` where :math:`N` is the number of nodes (vertices)
             and :math:`D` is the dimension of the embedding. Each row corresponds to the coordinates
-            of a node in the order given by `node_list`.
+            of a node in the order given by :attr:`node_list`.
 
         Notes:
             - If no nodes have been added, returns an empty array of shape :math:`(0, 0)`.
             - The coordinate matrix is updated automatically as nodes are added or modified.
-            - The order of rows matches the order of node identifiers in `node_list`.
+            - The order of rows matches the order of node identifiers in :attr:`node_list`.
         """
         if self._coord_matrix is None:
             return np.empty((0, 0))
@@ -105,7 +91,7 @@ class EmbeddedComplex(nx.Graph):
     @property
     def dim(self):
         """
-        Get the dimension of the embedded coordinates.
+        Get the dimension :math:`D` of the embedded coordinates.
 
         Returns:
             int: The number of dimensions (D) for the coordinates of each node. Returns 0 if no nodes exist.
@@ -120,14 +106,14 @@ class EmbeddedComplex(nx.Graph):
         Get the ordered list of node names in the complex.
 
         Returns:
-            list: List of node identifiers in the order corresponding to the rows of the coordinate matrix.
+            list: List of node identifiers in the order corresponding to the rows of the coordinate matrix :attr:`coord_matrix`.
         """
         return self._node_list
 
     @property
     def node_to_index(self):
         """
-        Get a mapping from node identifiers to their row index in the coordinate matrix.
+        Get a mapping from node identifiers to their row index in the coordinate matrix :attr:`coord_matrix`.
 
         Returns:
             dict: Dictionary mapping node ids to their index in the coordinate matrix.
[…]
