+++ src/ect/results.py
@@ -64,11 +64,36 @@ def _dense_to_csr_prefix(dense64: np.ndarray):
 
 class ECTResult(np.ndarray):
     """
-    A numpy ndarray subclass that carries ECT metadata and plotting capabilities
-    Acts like a regular matrix but with added visualization methods and metadata about directions and thresholds
+    A numpy ndarray subclass that carries Euler Characteristic Transform (ECT) metadata and plotting capabilities.
+
+    This class acts like a regular matrix but includes:
+    - Visualization methods for ECT and ECC plots
+    - Metadata about directions and thresholds used in the transform
+    - Support for both dense and compressed (CSR) representations
+
+    Parameters:
+        matrix (np.ndarray): ECT matrix data.
+        directions: Direction metadata (angles or vectors).
+        thresholds: Threshold values used for the transform.
+
+    Notes:
+        - Use :meth:`plot` to visualize the ECT matrix.
+        - Use :meth:`dist` to compute distances between ECT results.
+        - Supports loading/saving in compressed format via :meth:`save_npz` and :meth:`load_npz`.
     """
 
     def __new__(cls, matrix, directions, thresholds):
+        """
+        Create a new ECTResult instance from matrix data, directions, and thresholds.
+
+        Args:
+            matrix (np.ndarray): ECT matrix data (float or int).
+            directions: Direction metadata (angles or vectors).
+            thresholds: Threshold values used for the transform.
+
+        Returns:
+            ECTResult: A new instance with attached metadata and optional CSR fields.
+        """
         # allow float arrays for smooth transform otherwise int
         if np.issubdtype(matrix.dtype, np.floating):
             obj = np.asarray(matrix, dtype=np.float64).view(cls)
@@ -92,6 +117,12 @@ class ECTResult(np.ndarray):
 
     @property
     def has_csr(self):
+        """
+        Check if the ECTResult instance has compressed sparse row (CSR) data attached.
+
+        Returns:
+            bool: True if CSR fields (row_ptr, col_idx, data) are present, False otherwise.
+        """
         return (
             getattr(self, "csr_row_ptr", None) is not None
             and getattr(self, "csr_col_idx", None) is not None
@@ -100,6 +131,20 @@ class ECTResult(np.ndarray):
 
     @classmethod
     def from_csr(cls, row_ptr, col_idx, data, directions, thresholds, dtype=np.int32):
+        """
+        Create an ECTResult instance from compressed sparse row (CSR) data.
+
+        Args:
+            row_ptr (np.ndarray): CSR row pointer array.
+            col_idx (np.ndarray): CSR column indices array.
+            data (np.ndarray): CSR data array (jump values).
+            directions: Direction metadata (angles or vectors).
+            thresholds: Threshold values used for the transform.
+            dtype: Data type for the dense matrix (default: np.int32).
+
+        Returns:
+            ECTResult: Instance with dense matrix and attached CSR fields.
+        """
         num_dirs = len(directions)
         num_thresh = len(thresholds)
         dense64 = _csr_prefix_to_dense(row_ptr, col_idx, data, num_dirs, num_thresh)
@@ -111,6 +156,12 @@ class ECTResult(np.ndarray):
         return obj
 
     def to_dense(self):
+        """
+        Convert the ECTResult from CSR format to a dense matrix.
+
+        Returns:
+            np.ndarray: Dense matrix representation of the ECTResult.
+        """
         if not self.has_csr:
             return self
         num_dirs = self.shape[0]
@@ -121,6 +172,16 @@ class ECTResult(np.ndarray):
         return dense64.astype(self.dtype, copy=False)
 
     def save_npz(self, path):
+        """
[…]
