+++ src/ect/ect.py
@@ -16,16 +16,7 @@ class ECT:
     The result is a matrix where entry ``M[i,j]`` is :math:`\chi(K_{a_i})` for the direction :math:`\omega_j`
     where :math:`a_i` is the ith entry in ``self.thresholds``, and :math:`\omega_j` is the jth entry in ``self.directions``.
 
-    Attributes
-    ----------
-    num_dirs : int
-        The number of directions to consider in the matrix.
-    num_thresh : int
-        The number of thresholds to consider in the matrix.
-    directions : Directions
-        The directions to consider for projection.
-    bound_radius : float
-        Either ``None``, or a positive radius of the bounding circle.
+    
 
     Example:
         >>> from ect import ECT, EmbeddedComplex
