### 2026-01-29 5b9c08c
Add test for notebook tutorial execution
