### 2026-01-29 f6a473f
Update version to 1.2.3 and add nbformat and nbclient
