### 2026-01-29 d7c2bc7
remove notebook specific tests
