### 2026-01-29 67d9e8e
Update documentation configuration and workflow settings for documentation testing
