+++ doc_source/conf.py
@@ -12,10 +12,13 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.join(os.path.dirname(__file__),'..'))
-sys.path.insert(0, os.path.join(os.path.dirname(__file__),'..', 'ect' ))
-sys.path.insert(0, os.path.join(os.path.dirname(__file__),'..', 'ect', 'ect_on_graphs'))
+
+sys.path.insert(0, os.path.abspath("."))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "ect"))
+sys.path.insert(
+    0, os.path.join(os.path.dirname(__file__), "..", "ect", "ect_on_graphs")
+)
 # sys.path.insert(0, os.path.join(os.path.dirname(__file__),'..', 'teaspoon'))
 # sys.path.insert(0, os.path.join(os.path.dirname(__file__)))
 # sys.path.insert(0, os.path.join(os.path.dirname(__file__),'..','..'))
@@ -23,12 +26,12 @@ sys.path.insert(0, os.path.join(os.path.dirname(__file__),'..', 'ect', 'ect_on_g
 
 # -- Project information -----------------------------------------------------
 
-project = 'ect'
-copyright = '2025, MunchLab'
-author = 'MunchLab'
+project = "ect"
+copyright = "2025, MunchLab"
+author = "MunchLab"
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.5'
+release = "0.1.5"
 
 
 # -- General configuration ---------------------------------------------------
@@ -36,27 +39,29 @@ release = '0.1.5'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc',
-            #   'numpydoc',
-              'sphinx.ext.coverage',
-              'sphinx.ext.napoleon',
-              'matplotlib.sphinxext.mathmpl',
-              'matplotlib.sphinxext.plot_directive',
-              'sphinx.ext.autosummary',
-              'sphinx.ext.doctest',
-              'sphinx.ext.intersphinx',
-              'sphinx.ext.todo',
-              'sphinx.ext.mathjax',
-              'sphinx.ext.ifconfig',
-              'sphinx.ext.viewcode',
-              'sphinx.ext.githubpages',
-              'myst_parser', 
-              'sphinx.ext.autosummary',
-              'nbsphinx',
-              'sphinx.ext.viewcode']
+extensions = [
+    "sphinx.ext.autodoc",
+    #   'numpydoc',
+    "sphinx.ext.coverage",
+    "sphinx.ext.napoleon",
+    "matplotlib.sphinxext.mathmpl",
+    "matplotlib.sphinxext.plot_directive",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.doctest",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.todo",
+    "sphinx.ext.mathjax",
+    "sphinx.ext.ifconfig",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.githubpages",
+    "myst_parser",
+    "sphinx.ext.autosummary",
+    "nbsphinx",
+    "sphinx.ext.viewcode",
+]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -69,30 +74,30 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
[…]
