+++ .github/workflows/docs.yml
@@ -1,15 +1,12 @@
 name: Build and Deploy Docs
 on:
   push:
-    branches: [ "main" ]
+    branches: ["main"]
     paths:
-      - 'doc_source/**'
-      - 'src/ect/**'
+      - "doc_source/**"
+      - "src/ect/**"
   pull_request:
-    branches: [ "main" ]
-    paths:
-      - 'doc_source/**'
-      - 'src/ect/**'
+    branches: ["main"]
 
 jobs:
   docs:
@@ -20,36 +17,36 @@ jobs:
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
-          
+
       - name: Install uv
         uses: astral-sh/setup-uv@v5
-        
+
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
           python-version-file: "pyproject.toml"
-          
+
       - name: Install system dependencies
         run: |
           sudo apt-get update
           sudo apt-get install -y pandoc
-          
+
       - name: Create venv and install dependencies
         run: |
           uv venv
           source .venv/bin/activate
           uv pip install -e .
           uv pip install sphinx sphinx-rtd-theme nbsphinx autopep8 myst-parser
-          
+
       - name: Build documentation
         run: |
           source .venv/bin/activate
           make html
-          
+
       - name: Deploy to GitHub Pages
         if: github.event_name == 'push'
         uses: peaceiris/actions-gh-pages@v3
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: ./docs
-          force_orphan: true 
+          force_orphan: true
