### 2026-01-29 13f294f
Add installation of python3 Jupyter kernel
