+++ .github/workflows/docs.yml
@@ -38,6 +38,11 @@ jobs:
           uv pip install -e .
           uv pip install sphinx sphinx-rtd-theme nbsphinx autopep8 myst-parser
 
+      - name: Install python3 Jupyter kernel
+        run: |
+          source .venv/bin/activate
+          python -m ipykernel install --user --name python3 --display-name "Python 3"
+
       - name: Build documentation
         run: |
           source .venv/bin/activate
