### 2026-01-29 5734f57
Add ipykernel to the list of installed packages
