### 2026-01-29 b3abfa8
Add pandas library to the documentation workflow
