### 2026-01-29 c9de1f0
add sect and dect citations
