### 2026-01-29 d7f323e
update in text citations correctly and fix typos
