+++ _sources/tutorials.rst.txt
@@ -1,0 +1,10 @@
+Tutorials
+=========
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   notebooks/Tutorial-EmbeddedComplex
+   notebooks/Matisse/example_matisse
+
