+++ _sources/installation.rst.txt
@@ -1,0 +1,37 @@
+Installation
+============
+
+Prerequisites
+-------------
+
+Before installing `ect`, make sure you have the following prerequisites:
+
+- Python (version 3.10 or higher)
+- Pip (Python package installer)
+
+Installing `ect`
+----------------
+
+To install `ect`, run the following in a terminal.
+
+.. code-block:: bash
+
+    pip install ect
+
+Alternatively, you can install directly from the source by cloning the repository and running the following command:
+
+.. code-block:: bash
+
+    git clone https://github.com/MunchLab/ect
+    cd ect
+    pip install .
+
+
+Uninstalling `ect`
+------------------
+
+To uninstall `ect`, simply run the following command:
+
+.. code-block:: bash
+
+    pip uninstall ect
