+++ _sources/citing.rst.txt
@@ -1,0 +1,7 @@
+Citing
+=======
+
+To cite `ect` please use the following publication:
+
+   Munch, Elizabeth. An Invitation to the Euler Characteristic Transform. The American Mathematical Monthly, 132(1), 15-25. `doi:10.1080/00029890.2024.2409616 <https://doi.org/10.1080/00029890.2024.2409616>`_. 2024.
+
