### 2026-02-08 #9691 @raphaelreinauer
@prashjha Thanks for the ping. I'm standing by to finalize my checklist, but I am waiting on the authors to address the issues I raised in my initial review ([#80](https://github.com/MunchLab/ect/issues/86)–[#86](https://github.com/MunchLab/ect/issues/86)).

@yemeen Just checking in on these, do you need any clarification from me on any of the iussues? Happy to discuss them in the repo if needed.
