### 2026-02-19 94f9a2f
Ensure uniform thresholds validation in ECT calculation
