+++ paper/paper.bib
@@ -191,4 +191,68 @@
   publisher = {Center for Open Science},
 }
 
+@Article{Tang2022,
+  author    = {Wai Shing Tang and Gabriel Monteiro da Silva and Henry Kirveslahti and Erin Skeens and Bibo Feng and Timothy Sudijono and Kevin K. Yang and Sayan Mukherjee and Brenda Rubenstein and Lorin Crawford},
+  journal   = {{PLOS} Computational Biology},
+  title     = {A topological data analytic approach for discovering biophysical signatures in protein dynamics},
+  year      = {2022},
+  month     = {may},
+  number    = {5},
+  pages     = {e1010045},
+  volume    = {18},
+  doi       = {10.1371/journal.pcbi.1010045},
+  editor    = {Petras J Kundrotas},
+  file      = {:DirectionalTransform/Tang2022.pdf:PDF},
+  publisher = {Public Library of Science ({PLoS})},
+}
+
+@Article{Amezquita2021,
+  author    = {Erik J Am{\'{e}}zquita and Michelle Y Quigley and Tim Ophelders and Jacob B Landis and Daniel Koenig and Elizabeth Munch and Daniel H Chitwood},
+  journal   = {in silico Plants},
+  title     = {Measuring hidden phenotype: quantifying the shape of barley seeds using the Euler characteristic transform},
+  year      = {2021},
+  month     = {dec},
+  number    = {1},
+  volume    = {4},
+  doi       = {10.1093/insilicoplants/diab033},
+  editor    = {Tsu-Wei Chen and Stephen P Long},
+  file      = {:DirectionalTransform/Amezquita2021a.pdf:PDF},
+  publisher = {Oxford University Press ({OUP})},
+}
+
+@Misc{CisewskiKehe2025,
+  author        = {Jessi Cisewski-Kehe and Brittany Terese Fasy and Alexander McCleary and Eli Quist and Jack Ruder},
+  title         = {Vectorized Computation of Euler Characteristic Functions and Transforms},
+  year          = {2025},
+  archiveprefix = {arXiv},
+  eprint        = {2511.03909},
+  primaryclass  = {cs.CG},
+  url           = {https://arxiv.org/abs/2511.03909},
+}
+
+@Article{CisewskiKehe2023,
+  author        = {Cisewski-Kehe, Jessi and Fasy, Brittany Terese and Giriyan, Dhanush and Quist, Eli},
+  title         = {The Weighted Euler Characteristic Transform for Image Shape Classification},
+  year          = {2023},
+  month         = jul,
+  abstract      = {The weighted Euler characteristic transform (WECT) is a new tool for extracting shape information from data equipped with a weight function. Image data may benefit from the WECT where the intensity of the pixels are used to define the weight function. In this work, an empirical assessment of the WECT's ability to distinguish shapes on images with different pixel intensity distributions is considered, along with visualization techniques to improve the intuition and understanding of what is captured by the WECT. Additionally, the expected weighted Euler characteristic and the expected WECT are derived.},
+  archiveprefix = {arXiv},
+  copyright     = {arXiv.org perpetual, non-exclusive license},
+  doi           = {10.48550/ARXIV.2307.13940},
+  eprint        = {2307.13940},
+  file          = {:http\://arxiv.org/pdf/2307.13940v1:PDF;:DirectionalTransform/CisewskiKehe2023.pdf:PDF},
+  keywords      = {Computational Geometry (cs.CG), Algebraic Topology (math.AT), FOS: Computer and information sciences, FOS: Mathematics},
+  primaryclass  = {cs.CG},
+  publisher     = {arXiv},
+}
+
[…]
