+++ paper/paper.bib
@@ -255,4 +255,19 @@
   file      = {:DirectionalTransform/Jiang2020.pdf:PDF},
 }
 
+@Article{PlantsAndPython2022,
+  journal  = {The Plant Cell},
+  title    = {Plants \&amp; Python: A series of lessons in coding, plant biology, computation, and bioinformatics},
+  year     = {2022},
+  issn     = {1040-4651},
+  month    = {07},
+  number   = {7},
+  pages    = {e1-e1},
+  volume   = {34},
+  abstract = {(By Robert VanBuren, Alejandra Rougon-Cardoso, Erik J. Amézquita, Evelia L. Coss-Navarrete, Aarón Espinosa-Jaime, Omar Andres Gonzalez-Iturbe, Alicia C. Luckie-Duque, Eddy Mendoza-Galindo, Jeremy Pardo, Guillermo Rodríguez-Guerrero, Pablo Y. Rosiles-Loeza, Marilyn Vásquez-Cruz, Selene L. Fernandez-Valverde, Tania Hernández-Hernández, Sourabh Palande, and Daniel H. Chitwood)},
+  doi      = {10.1093/plcell/koac187},
+  eprint   = {https://academic.oup.com/plcell/article-pdf/34/7/e1/44459508/koac187.pdf},
+  url      = {https://doi.org/10.1093/plcell/koac187},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
