### 2026-02-25 82925c7
Add benchmarking code for ECT package implementations and edit paper
