+++ benchmarks/benchmark_ect_packages.py
@@ -1,0 +1,777 @@
+#!/usr/bin/env python3
+import time
+
+import matplotlib.pyplot as plt
+import numpy as np
+import torch
+from dect.ect import compute_ect as DECT_COMPUTE_ECT
+from dect.ect_fn import scaled_sigmoid as DECT_SCALED_SIGMOID
+from pyect import WECT as PYECT_WECT
+
+from ect import ECT, EmbeddedComplex, Directions
+
+N_POINTS = 500
+DIM = 3
+NUM_DIRS = 256
+NUM_THRESH = 256
+RESOLUTIONS = "32,64,128,256,512"
+REPEATS = 10
+SEED = 42
+DECT_SCALE = 500.0
+STRUCTURE = "grid2d"
+TORCH_DEVICE = "cpu"
+
+
+def generate_points(n_points: int, dim: int, seed: int) -> np.ndarray:
+    """
+    Generate random points in [0,1]^dim.
+    """
+    rng = np.random.default_rng(seed)
+    return rng.random((n_points, dim), dtype=np.float64)
+
+
+def build_point_only_complex(points: np.ndarray) -> EmbeddedComplex:
+    """
+    Construct an EmbeddedComplex containing only 0-cells (vertices).
+    This ensures our ECT matches a constant-weight WECT computed on the same set of points.
+    """
+    K = EmbeddedComplex()
+    nodes_with_coords = [(i, points[i]) for i in range(points.shape[0])]
+    K.add_nodes_from(nodes_with_coords)
+    return K
+
+
+def build_grid2d_complex(grid_size: int) -> EmbeddedComplex:
+    """
+    Build a simple 2D triangular grid complex with vertices, edges, and triangular faces.
+    Vertices are on a unit square grid of shape (grid_size x grid_size).
+    """
+    K = EmbeddedComplex()
+    m = grid_size
+    # Add vertices with coordinates
+    coords = []
+    for y in range(m):
+        for x in range(m):
+            coords.append([x / (m - 1), y / (m - 1)])
+    coords = np.asarray(coords, dtype=np.float64)
+    K.add_nodes_from([(i, coords[i]) for i in range(coords.shape[0])])
+
+    def vid(x, y):
+        return y * m + x
+
+    # Add grid edges (right and down neighbors)
+    edges = []
+    for y in range(m):
+        for x in range(m):
+            if x + 1 < m:
+                edges.append((vid(x, y), vid(x + 1, y)))
+            if y + 1 < m:
+                edges.append((vid(x, y), vid(x, y + 1)))
+    if edges:
+        K.add_edges_from(edges)
+
+    # Add faces: split each cell into two triangles
+    for y in range(m - 1):
+        for x in range(m - 1):
+            v00 = vid(x, y)
+            v10 = vid(x + 1, y)
+            v01 = vid(x, y + 1)
+            v11 = vid(x + 1, y + 1)
+            K.add_cell([v00, v10, v11], dim=2)
+            K.add_cell([v00, v11, v01], dim=2)
+
+    return K
+
+
+def _shared_directions(num_dirs: int, dim: int, seed: int) -> np.ndarray:
+    if dim == 2:
+        angles = np.linspace(
+            0.0, 2.0 * np.pi, num_dirs, endpoint=False, dtype=np.float64
+        )
+        return np.stack([np.cos(angles), np.sin(angles)], axis=1)
+    rng = np.random.default_rng(seed)
+    dirs = rng.standard_normal((num_dirs, dim), dtype=np.float64)
+    norms = np.linalg.norm(dirs, axis=1, keepdims=True)
+    norms = np.maximum(norms, 1e-12)
+    return dirs / norms
+
+
[…]
