### 2026-02-25 5fdef1a
Update software implementations for ECT and its variants
