### 2026-02-25 0fb973f
Update data types to float for direction angles and vectors
