+++ src/ect/directions.py
@@ -16,7 +16,7 @@ class Directions:
     Supports uniform, random, or custom sampling of directions.
 
     **Examples:**
-    
+
     .. code-block:: python
 
         # Uniform sampling in 2D (default)
@@ -57,15 +57,15 @@ class Directions:
             - For 2D, directions are represented as angles; for higher dimensions, as unit vectors.
             - Use factory methods :meth:`uniform`, :meth:`random`, :meth:`from_angles`, or :meth:`from_vectors` for convenience.
         """
-        
+
         self.num_dirs = num_dirs
         self.sampling = sampling
         self.dim = dim
         self.endpoint = endpoint
 
         self._rng = np.random.RandomState(seed)
-        self._thetas = None
-        self._vectors = None
+        self._thetas: Optional[np.ndarray] = None
+        self._vectors: Optional[np.ndarray] = None
         self._initialize_directions()
 
     def _initialize_directions(self):
@@ -82,16 +82,14 @@ class Directions:
             else:
                 # generate random normal samples and normalize to lie on the unit sphere
                 self._vectors = self._rng.randn(self.num_dirs, self.dim)
-                self._vectors /= np.linalg.norm(self._vectors,
-                                                axis=1, keepdims=True)
+                self._vectors /= np.linalg.norm(self._vectors, axis=1, keepdims=True)
         elif self.sampling == Sampling.RANDOM:
             if self.dim == 2:
                 self._thetas = self._rng.uniform(0, 2 * np.pi, self.num_dirs)
                 self._thetas.sort()
             else:
                 self._vectors = self._rng.randn(self.num_dirs, self.dim)
-                self._vectors /= np.linalg.norm(self._vectors,
-                                                axis=1, keepdims=True)
+                self._vectors /= np.linalg.norm(self._vectors, axis=1, keepdims=True)
 
     @classmethod
     def uniform(
@@ -142,11 +140,11 @@ class Directions:
             - Angles are stored in :attr:`thetas` and unit vectors are computed as needed.
         """
         instance = cls(len(angles), Sampling.CUSTOM, dim=2)
-        instance._thetas = np.array(angles)
+        instance._thetas = np.array(angles, dtype=float)
         return instance
 
     @classmethod
-    def from_vectors(cls, vectors: Sequence[tuple]) -> "Directions":
+    def from_vectors(cls, vectors: Sequence[Sequence[float]]) -> "Directions":
         """
         Create a Directions instance from custom direction vectors in any dimension.
 
@@ -163,12 +161,12 @@ class Directions:
             - Vectors are normalized to unit length.
             - For 2D, angles are computed from the vectors and available via :attr:`thetas`.
         """
-        vectors = np.array(vectors, dtype=float)
-        norms = np.linalg.norm(vectors, axis=1, keepdims=True)
+        vectors_arr = np.array(vectors, dtype=float)
+        norms = np.linalg.norm(vectors_arr, axis=1, keepdims=True)
         if np.any(norms == 0):
             raise ValueError("Zero-magnitude vectors are not allowed")
-        normalized = vectors / norms
-        instance = cls(len(vectors), Sampling.CUSTOM, dim=vectors.shape[1])
+        normalized = vectors_arr / norms
+        instance = cls(len(vectors_arr), Sampling.CUSTOM, dim=vectors_arr.shape[1])
         instance._vectors = normalized
         if instance.dim == 2:
             instance._thetas = np.arctan2(normalized[:, 1], normalized[:, 0])
@@ -196,6 +194,7 @@ class Directions:
         if self._thetas is None:
             # Compute the angles from the vectors.
             self._thetas = np.arctan2(self.vectors[:, 1], self.vectors[:, 0])
+        assert self._thetas is not None
         return self._thetas
 
     @property
@@ -215,13 +214,15 @@ class Directions:
[…]
