### 2026-02-26 #9691 @yemeen
hi everyone! we've added some changes to hopefully address the issues brought up in the reviews (state of the field, target audience, small code discrepancies). this includes adding comparisons to other packages, let me know if you consider them to be adequate (:

@shaelebrown @corybrunson @raphaelreinauer
