+++ paper/paper.bib
@@ -172,26 +172,26 @@
   publisher     = {arXiv}
 }
 
-@article{Yahiaoui2026,
+@Article{Yahiaoui2026,
   author    = {Yahiaoui, Wahiba and Smail, Safa and Ayub, Yemeen and Lu, Qiao and Cousins, Peter and Diaz-Garcia, Luis and Frank, Margaret and Headland, Lauren and Martinez, Ciera and Migicovsky, Zoë and Ranjan, Aashish and Sinha, Neelima and Swift, Joel F. and Torres-Lomas, Efrain and Munch, Elizabeth and Laiadi, Ziane and Chitwood, Daniel H.},
   title     = {Disentangling blade and vasculature shape in grapevine leaves},
   year      = {2026},
   month     = jan,
   doi       = {10.64898/2026.01.27.701982},
   keywords  = {inSubmission},
-  publisher = {openRxiv}
+  publisher = {openRxiv},
 }
 
-@article{GarciaChavez2026,
+@Article{GarciaChavez2026,
   author    = {García-Chávez, J. Noé and Vicente-Ferrer, Ángel and Bucio, Andrés Tapia and Ruíz-Amaro, Mariana J. and Hurtado-Olvera, J. Jimena and DeViller, Kylie and Simms, Lachlann and Ayub, Yemeen and Lu, Qiao and Anandappa, Jason Anthony and Bautista-Hinojosa, Luis and Boren, Duncan M. and Buitrago, Evelin and Buitrago-Acosta, Maria Camila and CG-Luna, Coral and Chalise, Devendra Prasad and Chamkasem, Aekkapot and Chennuru, Vineeth and Corpus-Hernández, Marcela and Espíndola, Brian S. and Gasdick, Michael C. and Gedela, Nitish Satya Sai and Guzmán-Barrenechea, Karla and Heshmati, Siavash and Jiménez-Rabadán, N. Gabriela and Jing, Vientiane and Juárez-Mercado, K. Eurídice and Konda, Sai Divya Teja and Korkmaz, Ilayda and Lakshmi-Narayanan, Aditya and Linn, Christina M. and Liu, Tiancheng and Lizardo, Viridiana and Lovanxay, Phanutda and Mahikul, Wiriya and Morales-Salazar, Alfredo and Morin, Ashlynn and Nguyen, Thanh Trung and Perkinson, Payton C. and Rojas-Roldán, Andrés and Sánchez-Fernández, Eduardo and Severiano-Galeana, Fernando and Uchima-Taborda, Diego and Zhu, Yuhan and Krishnan, Arjun and Husbands, Aman Y. and Romero-Hernandez, Annabel and VanBuren, Robert and Munch, Elizabeth and Migicovsky, Zoë and Rougon-Cardoso, Alejandra and Chitwood, Daniel H.},
   title     = {Indigenous Plant Knowledge in the De la Cruz-Badiano Codex (1552): A Text Mining and Morphometric Study},
   year      = {2026},
   month     = feb,
   doi       = {10.31235/osf.io/dmb9k_v1},
-  publisher = {Center for Open Science}
+  publisher = {Center for Open Science},
 }
 
-@article{Tang2022,
+@Article{Tang2022,
   author    = {Wai Shing Tang and Gabriel Monteiro da Silva and Henry Kirveslahti and Erin Skeens and Bibo Feng and Timothy Sudijono and Kevin K. Yang and Sayan Mukherjee and Brenda Rubenstein and Lorin Crawford},
   journal   = {{PLOS} Computational Biology},
   title     = {A topological data analytic approach for discovering biophysical signatures in protein dynamics},
@@ -203,10 +203,10 @@
   doi       = {10.1371/journal.pcbi.1010045},
   editor    = {Petras J Kundrotas},
[…]
