### 2026-02-26 3abf830
add dist matrix function and update tutorial
