### 2026-02-26 3e623ef
Bump version to 1.2.4 in pyproject.toml
