+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "ect"
-version = "1.2.3"
+version = "1.2.4"
 authors = [
   { name="Liz Munch", email="muncheli@msu.edu" },
 ]
