### 2026-02-27 MunchLab/ect#72 @corybrunson
@yemeen the new section surely satisfies this issue. A minor syntactic suggestion is to hyphenate "application-specific".
