+++ _sources/notebooks/Tutorial-EmbeddedComplex.ipynb.txt
@@ -1,0 +1,268 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Tutorial: ECT for Embedded Cell Complexes\n",
+    "\n",
+    "This tutorial will walk you through using the `ECT` package. Particularly we will show the features of `EmbeddedComplex` class for computing the Euler Characteristic Transform on complexes with arbitrary dimensional cells.\n",
+    "\n",
+    "The `EmbeddedComplex` class combines and extends the functionality of the previous `EmbeddedGraph` and `EmbeddedCW` classes, supporting **k-cells** for k ≥ 2.\n",
+    "\n"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [
+    {
+     "ename": "",
+     "evalue": "",
+     "output_type": "error",
+     "traceback": [
+      "\u001b[1;31mRunning cells with '.conda-ect-3-11 (Python 3.11.13)' requires the ipykernel package.\n",
+      "\u001b[1;31mInstall 'ipykernel' into the Python environment. \n",
+      "\u001b[1;31mCommand: 'conda install -n .conda-ect-3-11 ipykernel --update-deps --force-reinstall'"
+     ]
+    }
+   ],
+   "source": [
+    "import numpy as np\n",
+    "import matplotlib.pyplot as plt\n",
+    "from ect import EmbeddedComplex, ECT, Directions\n",
+    "from ect.utils.examples import create_example_graph, create_example_cw, create_example_3d_complex"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Basic Usage: Creating Simple Complexes\n",
+    "\n",
+    "### Example 1: Graph (1-skeleton)\n",
+    "\n",
+    "Let's start with a simple triangle graph (for legacy users this can be equivalently done using `EmbeddedGraph`). "
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "K = EmbeddedComplex()\n",
+    "\n",
+    "K.add_node('A', [0, 0])\n",
+    "K.add_node('B', [1, 0])\n",
+    "K.add_node('C', [0.5, 0.866])\n",
+    "\n",
+    "K.add_edge('A', 'B')\n",
+    "K.add_edge('B', 'C')\n",
+    "K.add_edge('C', 'A')\n",
+    "\n",
+    "#using built-in plotting function along with matplotlib\n",
+    "fig, ax = plt.subplots(figsize=(6, 6))\n",
+    "K.plot(ax=ax, with_labels=True, node_size=400)\n",
+    "ax.set_title('Simple Triangle Graph\\n(0-cells: vertices, 1-cells: edges)')\n",
+    "plt.show()\n",
+    "\n",
+    "#print some information about the complex\n",
+    "print(f\"Number of nodes: {len(K.nodes())}\")\n",
+    "print(f\"Number of edges: {len(K.edges())}\")\n",
+    "print(f\"Embedding dimension: {K.dim}\")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Now let's add a 2-cell (face) to fill in the triangle:"
+   ]
+  },
+  {
+   "cell_type": "code",
[…]
