### 2026-03-07 #9691 @corybrunson
@yemeen i've been unable to reproduce the figures in [Tutorial 3.2](https://munchlab.github.io/ect/notebooks/Matisse/example_matisse.html). This could easily be a problem with my understanding or implementation of Python. @shaelebrown or @raphaelreinauer, could you confirm that you have reproduced that document?
