### 2026-03-08 #9691 @raphaelreinauer
> [@yemeen](https://github.com/yemeen) i've been unable to reproduce the figures in [Tutorial 3.2](https://munchlab.github.io/ect/notebooks/Matisse/example_matisse.html). This could easily be a problem with my understanding or implementation of Python. [@shaelebrown](https://github.com/shaelebrown) or [@raphaelreinauer](https://github.com/raphaelreinauer), could you confirm that you have reproduced that document?

@corybrunson Yes, I was able to reproduce Tutorial 3.2 locally, including the final MDS plot.
