### 2026-03-08 #9691 @raphaelreinauer
@prashjha I went through issues MunchLab/ect#80-MunchLab/ect#86 one by one. From what I can tell, MunchLab/ect#83, MunchLab/ect#85, and MunchLab/ect#86 are addressed; MunchLab/ect#80 is mostly addressed via added benchmarking; MunchLab/ect#81 is only partially addressed; MunchLab/ect#82 still appears unresolved; and MunchLab/ect#84 still has at least one remaining typo. I’m therefore not quite ready to finalize my checklist yet.
