### 2026-03-08 #9691 @corybrunson
> [@corybrunson](https://github.com/corybrunson) Yes, I was able to reproduce Tutorial 3.2 locally, including the final MDS plot.

Thanks @raphaelreinauer!  @prashjha my checklist is complete.
