### 2026-03-08 #9691 @corybrunson
> [@prashjha](https://github.com/prashjha) I went through issues [#80](https://github.com/openjournals/joss-reviews/issues/80)-[#86](https://github.com/openjournals/joss-reviews/issues/86) one by one.

FYI @raphaelreinauer the autolinks are sending me to different `joss-reviews` threads.
