### 2026-03-12 #9691 @yemeen
@raphaelreinauer the changes for issues 81 and 82 had been merged into the JOSS-Paper branch on accident. this has been fixed now! i also believe i fixed the last typo.
