### 2026-03-12 #9691 @raphaelreinauer
> > [@corybrunson](https://github.com/corybrunson) Yes, I was able to reproduce Tutorial 3.2 locally, including the final MDS plot.
>
> Thanks [@raphaelreinauer](https://github.com/raphaelreinauer)! [@prashjha](https://github.com/prashjha) my checklist is complete.

Thanks @corybrunson, fixed it.
