### 2026-03-12 #9691 @raphaelreinauer
@yemeen Thanks, I checked again. MunchLab/ect#82 looks fixed on main as well as the typo. However, MunchLab/ect#81 still does not look fully resolved to me: while the ruff issues appear fixed, `mypy src/ect --ignore-missing-imports` still shows 14 errors on main.
More generally, checks like ruff and mypy should ideally run automatically in CI so these regressions are caught before merge.
