### 2026-03-12 2b892b9
Add type hinting for `all_coords` parameter in validation functions and fix the rest of mypy errors
