### 2026-03-12 #9691 @yemeen
@raphaelreinauer sorry about that mypy should now pass!
