### 2026-03-13 #9691 @raphaelreinauer
> @raphaelreinauer sorry about that mypy should now pass!

Thanks for letting me know. For my review process, I need to see the actual code changes rather than just verifying the output on main. Could you please link the PR that resolved this?

Going forward, it would be super helpful if you could link the relevant PRs or commits whenever an issue is solved, as well as having one PR per issue.
