### 2026-03-13 #9691 @yemeen
@raphaelreinauer of course, the mypy issues were resolved [here](https://github.com/MunchLab/ect/pull/99).
