### 2026-04-01 #9691 @prashjha
Hi @lizliz, some tasks for you:

1. Please archive the code release on Zenodo (if not already done) and provide the archive reference so I can associate it with your JOSS submission. Also, please ensure that the title and author details of the Zenodo archive match those of this JOSS article. Also, the archive should contain the version number associated with this JOSS article.

2. If you have released a new version of the code and would like to update the version here, let me know.

I will review and share the suggestions directly here or through a pull request soon.
