### 2026-04-01 7a637d7
Add threshold validation and search support for ECT calculation
