### 2026-04-03 c176fc8
Update package description for `ect` to include broader applications
