### 2026-04-03 #9691 @yemeen
hi @prashjha, the zenodo archive can be found here (https://zenodo.org/records/19406764). we would also like to update to the newest version of the code, 1.3.0.

thank you for everything!!
