### 2026-04-20 #9691 @prashjha
> hi [@prashjha](https://github.com/prashjha), the zenodo archive can be found here (https://zenodo.org/records/19406764). we would also like to update to the newest version of the code, 1.3.0.
>
> thank you for everything!!

Hi @yemeen, please edit the title of your archive to match the JOSS article title.
