### 2026-04-20 569d524
Joss review: update doi
In relation to: openjournals/joss-reviews#9691 and [comment](https://github.com/openjournals/joss-reviews/issues/9691#issuecomment-4172279183)
