+++ paper/paper.bib
@@ -253,6 +253,7 @@
   year      = {2020},
   month     = {June},
   file      = {:DirectionalTransform/Jiang2020.pdf:PDF},
+  doi       = {10.1109/cvprw50498.2020.00430},
 }
 
 @Article{PlantsAndPython2022,
