+++ paper/paper.bib
@@ -129,11 +129,16 @@
 }
 
 @book{Ghrist2014,
-  title    = {Elementary Applied Topology},
-  year     = {2014},
-  author   = {Robert Ghrist},
   keywords = {textbook, survey}
+  author = {Ghrist, Robert},
+  isbn = {978-1502880857},
+  month = sep,
+  publisher = {Createspace},
+  title = {Elementary Applied Topology},
+  url = {https://www2.math.upenn.edu/~ghrist/notes.html},
+  year = 2014
 }
+
 
 @book{Goodman2018,
   editor    = {Jacob E. Goodman and Joseph O'Rourke and Csaba D. Tóth},
