### 2026-04-20 f4d9153
Joss review: minor edits
In relation to: openjournals/joss-reviews#9691.

Based on the current JOSS draft, some points addressed in these edits:

- L54–55: The sentence beginning “Since its popularity...” used “since” twice and was revised for clarity.
- L24 and elsewhere: Standard names for NumPy, SciPy, and PyTorch should be used consistently.
- L72–73: The sentence beginning “The ect package is focused on...” was revised slightly for readability.
- L101: “based in part upon work supported in part by” contained repetition of “in part”. “This work” may read better than “This material”.

Sugge
[…]
