### 2026-04-20 #9691 @prashjha
Hi @yemeen, I applied some edits in pull request MunchLab/ect#102. Feel free to accept, reject, or make additional changes as you see fit. If you want to make edits yourself, I recommend reading my commit comment [here](https://github.com/prashjha/ect/commit/f4d915372a177edd565009fc86d76d3550927dc5).

Once you proceed with the changes in the draft and update the Zenodo archive title to match the JOSS article title, I will recommend acceptance, and one of the EICs will take a final call. Feel free to ping me if you have any questions.
