### 2026-04-24 #9691 @danielskatz
👋 @yemeen - I've suggested some changes based on proofreading, in https://github.com/MunchLab/ect/pull/104 - please merge this, or let me know what you disagree with.
