### 2026-04-24 #9691 @danielskatz
@yemeen - sorry, I had a typo and added an extra character to a doi field in the bib file - can you replace `10.48550/arXiv.2511.03909c` with `10.48550/arXiv.2511.03909`?
