+++ paper/paper.bib
@@ -230,6 +230,7 @@
   title         = {Vectorized Computation of Euler Characteristic Functions and Transforms},
   year          = {2025},
   archiveprefix = {arXiv},
+  doi           = {10.48550/arXiv.2511.03909},
   eprint        = {2511.03909},
   primaryclass  = {cs.CG},
   url           = {https://arxiv.org/abs/2511.03909},
