+++ paper/paper.bib
@@ -234,7 +234,6 @@
   eprint        = {2511.03909},
   primaryclass  = {cs.CG},
   url           = {https://arxiv.org/abs/2511.03909},
-  doi           = {10.48550/arXiv.2511.03909c},
 }
 
 @Article{CisewskiKehe2023,
