### 2026-04-24 #9691 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
