### 2026-04-24 #9691 @danielskatz
Congratulations to @yemeen (Yemeen Ayub) and co-authors on your publication!!

JOSS completely depends on volunteers - we couldn't do this without you
